Example demonstrating that current Groebner basis implementation is wrong.
I used the noncommutative version of Buchberger's algorithm from the
article ``Noncommutative Grobner Bases, and Projective Resolutions''
by Ed Green.  It only considers overlaps of the form

   LeadingPath( f ) * c = b * LeadingPath( g ),

and not of the form

   LeadingPath( f ) = b * LeadingPath( g ) * c.

Then it is necessary to tip reduce the relations while building the
Groebner basis, but this is omitted in the algorithm presented in the
article, and therefore also in my code.

The algebra A4 defined in test.g demonstrates the problem.  The
elements A4.acdg and A4.aefg should be the same, but they are
considered different by the code.