Using empty matrices in representations.
Removed code which treats zero spaces in representations as a special
case.  Assume that a representation with zero spaces at some vertices
is created with appropriate empty matrices at the adjacent arrows.

Using MakeZeroMatrix instead of NullMat to make the zero matrices in
QuiverRepresentationByArrows, to make sure that we get empty matrices
when needed.