Started to rewrite quiver representations based on vector space categories.
For each algebra, there may be several different categories of
representations, based on different vector space categories.  One of
these is the "default" one -- this is based on either the category of
row spaces or the category of column spaces, based on whether the
quiver is a left or right quiver.