Changed filters for TwoSidedIdealByGenerators to avoid problems.
At the moment, there are two different implementations for ideals in
path algebras: one based on Groebner bases and one based on viewing
the ideal as a representation.

This commit is just a quick fix to avoid getting the second kind of
ideal when constructing a quotient kQ/I of a path algebra.  The real
solution would be to combine the two implementations of ideals.