Unified the two implementations of ideals.
We have two implementations of ideals:

* Two-sided ideals in path algebras are implemented via Groebner
  bases.

* Left, right or two-sided ideals in finite-dimensional quiver
  algebras are implemented via viewing them as modules.

In the case of a two-sided ideal in a finite-dimensional path algebra,
both implementations are applicable.

These two implementations are now defined in the same type of ideal
object.  This means that for a two-sided ideal in a finite-dimensional
path algebra, we get both a Groebner basis and a bimodule
corresponding to the ideal.