Interning of objects.
* Added functionality for "interning" objects, that is, making sure
  that (for the objects that are "interned") there does not exist
  multiple objects which are equal to each other.

* Activated interning for quivers and quiver algebras.

The reason for this is that we want to avoid the situation that two
copies of the same algebra have different module categories.  Say that
we create the same algebra twice:

  A1 := PathAlgebra( Rationals, LeftQuiver( "Q(3)[a:1->2,b:2->3]" ) );
  A2 := PathAlgebra( Rationals, LeftQuiver( "Q(3)[a:1->2,b:2->3]" ) );

Now A1 and A2 are equal, but (before this commit) not identical:

  gap> A1 = A2;
  true
  gap> IsIdenticalObj( A1, A2 );
  false

This means, for instance, that LeftModuleCategory( A1 ) and
LeftModuleCategory( A2 ) are different categories, even though they
should be the same.

Interning of algebras makes sure that A1 and A2 above become identical
objects, and thus their module categories are also identical.

Interning of quivers is not really necessary right now, but will be
when (if?) we create categories of quivers.