Allow quiver representations over any abelian category.
In some situations, it is useful to be able to construct a kind of
generalized quiver representation which does not consist of vector
spaces and linear transformations, as usual, but rather of objects and
morphisms from some abelian category.

For example, if A and B are algebras (given by "left" quivers), then
we would normally describe an A--B-bimodule by a representation over
the tensor algebra A tensor B^op.  However, we can also describe such
a bimodule by a representation over B^op which consists of left
A-modules and A-module homomorphisms.  Similarly, we could also
describe it by a representation over A which consists of right
B-modules and right B-module homomorphisms.

To make it possible to construct such generalized representations, I
have changed the requirement for the underlying category of a
representation category from IsVectorSpaceCategory to
IsAbelianCategory.

There are still many places where it is assumed that a representation
consists of vector spaces, so for the moment, one needs to be a little
careful with generalized representations.  We should probably try to
make a clear distinction between functions that only work for "normal"
representations and those that work for generalized representations as
well.