Changes in vector spaces.
Made some changes in the category of vector spaces to make it easier
to create vector spaces consisting of vectors which are not lists of
numbers, such as a vector space Hom(M,N) of module homomorphisms.

Combined the types for row and column vectors to a single type called
"standard vector".  This has the benefit that when we need to
automatically create a new vector space (such as when computing a
kernel or a cokernel), we do not need to choose between creating a row
space or a column space; we simply create a standard vector space.