Create families for algebras and algebra elements.
I originally thought that each quiver should have its own family, and
I reused the same family for every algebra over that quiver.  At some
point I changed my mind and made a single family

  FamilyOfQuivers

for all quivers, with an associated family

  FamilyOfPaths

for all paths.  The reuse of these families for algebras and algebra
elements was, however, retained.

An effect of this was that equality comparison of a quiver and an
algebra was not defined.  That is, there was no method to handle a
call

  Q = A

where Q is a quiver and A an algebra.  GAP has a built-in method for
the \= operation which returns false if the objects are in different
families.  But since our quivers and algebras where in the same
family, this did not apply.

This again led to problem with the interning mechanism, as described
in #16.  The Intern function relies on being able to compare objects
using the \= operation, and this breaks when we try to compare a
quiver and an algebra.  (It takes some time to discover this, because
the only times the Intern function actually needs to call \= is when a
hash collision occurs.)

To fix this problem, I made a separate family

  FamilyOfQuiverAlgebras

for algebras, with an associated family

  FamilyOfQuiverAlgebraElements

for algebra elements.  Now equality comparison using \= should work
for all quivers and quiver algebras, and the interning mechanism
should work correctly in all cases.

Fixes #16.