Remove hidden dependency on the Polycyclic package
We had used the function IsIdentityMat without realizing that this
function is defined in the package Polycyclic, and not in GAP itself.
We could have added Polycyclic as a dependency in PackageInfo.g, but
since we don't really need that package for anything else, I have
instead rewritten the code to avoid calling IsIdentityMat.

Fixes #19