Add layered representations.
Can use AsLayeredRepresentation to convert a representation over a
tensor algebra (A tensor B) to a "layered" representation: Either a
representation over A consisting of B-representations, or a
representation over B consisting of A-representations.

Will use this to implement Hom(M,N) with module structure when M
and/or N is a bimodule.