Change Replace operation
In some situations I need a function which takes a list and changes
the element at a certain index to something different.  I wanted to
name this function Replace, but already had a Replace which did
something slightly different: It took as an argument the object which
was to be replaced, instead of its index in the list.

I have changed the functionality like this:

* The existing Replace is renamed to ReplaceObj
* New operation Replace with the new functionality

I made new files replace.g{d,i} for these (the old one was in
side.g{d,i}, where it really didn't belong).