Implement \* for morphisms.
Implemented \* to mean the same as PreCompose for field category
morphisms (which includes linear transformations, representation
homomorphisms and module homomorphisms).  This also means that the GAP
builtin \^ operator now can be used to obtain powers of a morphism.

Had to make some changes to existing implementations of the \*
operator for product of a scalar and a morphism to avoid that these
interfere with the new functionality.