Use nice bases for subspaces of QPA vector spaces.
It is possible to call the GAP function Subspace on a QPA vector
space.  However, since this is a function and not an operation, we
can't control what it does.  Before this commit, the result would be a
somewhat useless vector space -- one could not, for instance, get a
basis for it.

I have now tried to use the GAP mechanism of "handling by nice bases"
to make these vector spaces more usable.

Consider the following code:

gap> V := StandardVectorSpace( Rationals, 3 );;
gap> S := Subspace( V, [ StandardVector( Rationals, [ 1, -1, 0 ] ) ] );;

After this commit, the vector space S has a basis and can perform
membership tests, so we can for example do the following:

gap> BasisVectors( Basis( S ) );
[ <vector [ 1, -1, 0 ]> ]
gap> StandardVector( Rationals, [ 1, 0, 0 ] ) in S;
false
gap> StandardVector( Rationals, [ 5, -5, 0 ] ) in S;
true