Add operation MapRepresentation for transforming representations
The basic form of MapRepresentation takes the following arguments:
* a quiver representation R
* a function vfun
* a function afun
* a quiver representation category cat
It creates a new quiver representation in the category cat by applying
the function vfun to every object in R and the function afun to every
morphism in R.  Thus, it «maps» R through the functions vfun and afun.

There are several other forms of the function with different argument
types.  I should write some documentation that explains what they do.