Add new versions of Hom for representations
We should have Hom operations which handle all meaningful combinations
of bimodules and one-sided modules, and produce either a vector space
or a module as a result.  Furthermore, we should have Hom available
both as a functor with either argument fixed, and as a bifunctor:

  Hom(M,-)  and  Hom(-,N)  and  Hom(-,-)

This commit is my attempt to implement most of the functionality
needed to achieve this.  So far, the implementation is only for
representations, and not modules.

The most general functionality is in the operation HomFunctor, which
takes an integer and two representation categories as arguments, and
produces a Hom bifunctor.  If one or both of the categories are based
on tensor product algebras, then the range of the functor is a quiver
representation category.  The integer argument must have value 1 or 2,
and says which tensor factor to hom over -- this corresponds to
hom-ing over the left or right module structure.