Remove duplicate of IsAcyclicQuiver
Commit 6306bea9 introduced a duplicate declaration and implementation
of the property IsAcyclicQuiver.  The new implementation seems to not
work properly:

gap> Q := LeftQuiver( "Q(3)[a:1->2,b:2->3]" );;
gap> IsAcyclicQuiver( Q );
false

I'm not sure why this implementation, which comes from QPA1, does not
work.  But since we already had a working implementation, I removed
the new one and moved the original implementation to where the new one
was placed, so that it is placed together with related functions.

I also removed the old declaration, keeping the new one instead.