Fix problem with inconsistent handling of representation categories
We allow the underlying "vector space category" of a quiver
representation category to be any abelian category (not necessarily a
category of vector spaces).  This is used in the code for Hom
functors, where we in some cases take the underlying category to be a
representation category.

However, some old code has survived which assumes that the underlying
category is always the same default category of vector spaces.  I have
now tried to fix this (not very elegantly, but so that it should at
least work).