Compute basis for hom space by calling BasisForHom
The function BasisForHom computes the basis for a hom space by making
a set of linear equations and letting GAP solve this system.  When I
wrote the code for Hom spaces, I tried to compute the basis in a
different way, by using the underlying CAP category of vector spaces.
However, this turned out to be much slower than BasisOfHom.  I have
therefore changed the Hom spaces so that they call BasisOfHom instead.

Related to issue #23.