Make TensorProductOfAlgebras more consistent
There are two versions of the operation TensorProductOfAlgebras -- one
which takes two algebras as separate arguments, and one which takes a
list of algebras:

  TensorProductOfAlgebras( A, B )
  TensorProductOfAlgebras( algebras )

These had different -- and slightly incompatible -- implementations.
If both algebras had at least two relations, then the results of the
two calls

  TensorProductOfAlgebras( A, B )

and

  TensorProductOfAlgebras( [ A, B ] )

would not compare as equal with respect to \=, because they would get
their relations ordered differently.

To make sure that these agree with each other, I have removed the
separate implementation of the two-argument version, and replaced it
with a call to the one-argument version.