Make stable categories more general
We had implemented stable categories modulo projectives of quiver
representation categories.  However, nothing in the implementation
depended on the fact that the original category consisted of
representations (and not, say, modules).  By doing a few small changes
(mostly changing the filters of operations), I have now made it work
for module categories as well.