Reimplement Hom for modules
This is a new implementation of the functions

  Hom( side, M, N )
  HomFunctor( side, cat1, cat2 )

These functions are used to hom over some (not necessarily all) of the
structure of their arguments.  For instance, we can say

  Hom( LEFT, M, N )

where M is a left A-module and N is an A-B-bimodule, and the result
will be a right B-module consisting of the left A-homomorphisms from M
to N.  This result is called a "hom module", and lies in the filter
IsHomModule.

Previously, I had built this up from similar functions for
representations, which were implemented in a complicated and buggy
way.  I have now removed that layer, and tried to make the new
implementation more straightforward.

I have also started to make it possible to do something meaningful
with the elements of hom modules, that is, homomorphisms that go
between modules in different categories (such as from the left module
M to the bimodule N in the example above).  These morphisms are now
called "enriched quiver homomorphisms".