Add attribute HomSide for module homomorphisms
This attribute already exists for enriched module homomorphisms.  For
normal module homomorphisms, it gives the same result as Side, but
that is not the case for enriched module homomorphisms.  Adding this
attribute for normal module homomorphisms means that we can safely
call HomSide on a morphism even if we don't know whether it is a
normal or enriched module homomorphism.