Make some changes to module homomorphism constructors
Make sure that a quiver module homomorphism always gets the correct
source and range modules.  Earlier, we could determine a module
uniquely by the underlying representation and whether the module
should be left, right or bimodule.  But now we can also have
Hom-modules.  This means that we need to be more explicit about
exactly which modules we want to use as source and range when we
construct a module homomorphism.